rename yearinterval to years-interval and shortoptions to short-options
