+++ functions/slots/extensions/resolvers/short-options.js
